refactor: フォーム作成・編集画面のエディタ実装を共通化する
フォーム作成画面と編集画面で重複していた質問編集 UI と送信処理を共有化し、
質問種別ごとの選択肢編集や初期値生成を共通部品へ切り出して保守しやすくする。
